adapters/db/product.go
package db

import (
	"database/sql"
	"errors"

	_ "github.com/mattn/go-sqlite3" // sqlite3 driver
	"github.com/paulomalandrim/go-hexagonal/application"
)

type ProductDb struct {
	db *sql.DB
}

func NewProductDb(db *sql.DB) *ProductDb {
	return &ProductDb{db: db}
}

func (p *ProductDb) Get(id string) (application.ProductInterface, error) {

	if p.db == nil {
		return nil, errors.New("database is not initialized")
	}

	var product application.Product
	stmt, err := p.db.Prepare("select id, name, price, status from products where id = ?")
	if err != nil {
		return nil, err
	}
	err = stmt.QueryRow(id).Scan(&product.ID, &product.Name, &product.Price, &product.Status)
	if err != nil {
		return nil, err
	}
	return &product, nil
}
